Round coordinates fix design
