+++ app/src/main/res/layout/city_item.xml
@@ -1,7 +1,8 @@
 <?xml version="1.0" encoding="utf-8"?>
-<com.google.android.material.card.MaterialCardView xmlns:android="http://schemas.android.com/apk/res/android"
+<com.google.android.material.card.MaterialCardView
+    xmlns:android="http://schemas.android.com/apk/res/android"
     android:id="@+id/card"
-    android:layout_width="wrap_content"
+    android:layout_width="200dp"
     android:layout_height="wrap_content"
     android:layout_margin="4dp"
     android:backgroundTint="?attr/colorSecondary">
@@ -13,13 +14,17 @@
         android:orientation="vertical"
         android:padding="4dp">
 
-        <!-- Title, secondary and supporting text -->
+        <!-- Title  -->
         <TextView
             android:id="@+id/txt_city_name"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
+            android:ellipsize="end"
+            android:maxLines="1"
             android:text="@string/app_name"
             android:textAppearance="?attr/textAppearanceHeadline6" />
+
+        <!-- Secondary text -->
 
         <TextView
             android:id="@+id/txt_secondary_info"
@@ -30,12 +35,16 @@
             android:textAppearance="?attr/textAppearanceBody2"
             android:textColor="?android:attr/textColorSecondary" />
 
+        <!-- Supported text -->
+
         <TextView
             android:id="@+id/txt_support_info"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:layout_marginTop="4dp"
             android:text="@string/app_name"
+            android:maxLines="1"
+            android:ellipsize="end"
             android:textAppearance="?attr/textAppearanceBody2"
             android:textColor="?android:attr/textColorSecondary" />
